+++ cli/sync.ts
@@ -390,7 +390,7 @@ async function compareDynFSElement(
         }
       }
       return o;
-    } else if (k.endsWith("flow.yaml")) {
+    } else if (k.endsWith(".app.yaml")) {
       const o: any = yamlParse(v);
       const o2 = o["policy"];
 
@@ -403,6 +403,8 @@ async function compareDynFSElement(
         }
       }
       return o;
+    } else {
+      return yamlParse(v);
     }
   }
   for (const [k, v] of Object.entries(m1)) {
@@ -497,13 +499,10 @@ export async function ignoreF(wmillconf: {
 }): Promise<(p: string, isDirectory: boolean) => boolean> {
   let whitelist: { approve(file: string): boolean } | undefined = undefined;
 
-  if (wmillconf?.includes || wmillconf?.excludes) {
-    if (wmillconf?.excludes && !Array.isArray(wmillconf.includes)) {
-      throw new Error("wmill.yaml/includes must be an array");
-    }
-    if (wmillconf?.excludes && !Array.isArray(wmillconf.excludes)) {
-      throw new Error("wmill.yaml/excludes must be an array");
-    }
+  if (
+    (Array.isArray(wmillconf?.includes) && wmillconf?.includes?.length > 0) ||
+    (Array.isArray(wmillconf?.excludes) && wmillconf?.excludes?.length > 0)
+  ) {
     whitelist = {
       approve(file: string): boolean {
         return (
@@ -520,6 +519,7 @@ export async function ignoreF(wmillconf: {
         denies(file: string): boolean;
       }
     | undefined = undefined;
+
   try {
     const ignoreContent = await Deno.readTextFile(".wmillignore");
     const condensed = ignoreContent
@@ -535,7 +535,10 @@ export async function ignoreF(wmillconf: {
   } catch {}
 
   if (ign && whitelist) {
-    throw new Error("Cannot have both .wmillignore and wmill.yaml/includes");
+    log.error(
+      "Cannot have both .wmillignore and wmill.yaml/includes or excludes, ignoring .wmillignore"
+    );
+    ign = undefined;
   }
 
   // new Gitignore.default({ initialRules: ignoreContent.split("\n")}).ignoreContent).compile();
@@ -1035,12 +1038,12 @@ const command = new Command()
   .option("--skip-resources", "Skip syncing  resources")
   .option("--include-schedules", "Include syncing  schedules")
   .option(
-    "-i --includes <patterns...:file>",
-    "Patterns to specify which file to take into account (among files that are compatible with windmill). Patterns can include * (any string until '/') and ** (any string)"
+    "-i --includes <patterns:file[]>",
+    "Comma separated patterns to specify which file to take into account (among files that are compatible with windmill). Patterns can include * (any string until '/') and ** (any string)"
   )
   .option(
-    "-e --excludes <patterns...:file>",
-    "Patterns to specify which file to NOT take into account."
+    "-e --excludes <patterns:file[]>",
+    "Comma separated patterns to specify which file to NOT take into account."
   )
   // deno-lint-ignore no-explicit-any
   .action(pull as any)
@@ -1069,12 +1072,12 @@ const command = new Command()
   .option("--skip-resources", "Skip syncing  resources")
   .option("--include-schedules", "Include syncing  schedules")
   .option(
-    "-i --includes <patterns...:file>",
-    "Patterns to specify which file to take into account (among files that are compatible with windmill). Patterns can include * (any string until '/') and ** (any string)"
+    "-i --includes <patterns:file[]>",
+    "Comma separated patterns to specify which file to take into account (among files that are compatible with windmill). Patterns can include * (any string until '/') and ** (any string)"
   )
   .option(
-    "-e --excludes <patterns...:file>",
-    "Patterns to specify which file to NOT take into account."
+    "-e --excludes <patterns:file[]>",
+    "Comma separated patterns to specify which file to NOT take into account."
   )
   .option(
     "--message <message:string>",