+++ frontend/src/lib/components/EditableSchemaForm.svelte
@@ -102,7 +102,9 @@
 			const n = {}
 
 			;(schema.order as string[]).forEach((x) => {
-				n[x] = schema.properties[x]
+				if (schema.properties && schema.properties[x] != undefined) {
+					n[x] = schema.properties[x]
+				}
 			})
 
 			Object.keys(schema.properties ?? {})